BigDecimal to double

+++ src/main/java/com/labwork/interfaces/FunctionInterface.java
@@ -1,8 +1,5 @@
 package com.labwork.interfaces;
 
-import java.math.BigDecimal;
-
 public interface FunctionInterface {
-
-    BigDecimal calculate(final BigDecimal x, final BigDecimal delta);
+    double calculate(final double x, final double delta);
 }
